requirements.txt
numpy==1.26.4
Pillow==10.3.0
streamlit==1.35.0
